refactored and moved fragments to corresponding packages
