+++ app/src/main/AndroidManifest.xml
@@ -28,7 +28,7 @@
 
     <activity
       android:label="@string/app_name"
-      android:name=".MainActivity">
+      android:name=".controller.MainActivity">
     </activity>
   </application>
