--- app/src/main/java/edu/cnm/deepdive/picmegallery/controller/GalleryFragment.java
@@ -1,5 +1,0 @@
-package edu.cnm.deepdive.picmegallery.controller;
-
-public class GalleryFragment {
-
-}
