+++ app/src/main/res/layout/fragment_event_photos.xml
@@ -5,7 +5,7 @@
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context=".controller.EventPhotoFragment">
+  tools:context=".ui.eventsPhoto.EventPhotoFragment">
 
   <androidx.recyclerview.widget.RecyclerView
     android:id="@+id/event_photos"
