+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -30,13 +30,13 @@
   <fragment
     android:id="@+id/event_photos"
     android:label="@string/event_photos"
-    android:name="edu.cnm.deepdive.picmegallery.controller.EventPhotoFragment"
+    android:name="edu.cnm.deepdive.picmegallery.ui.eventsPhoto.EventPhotoFragment"
     tools:layout="@layout/fragment_event_photos"/>
 
   <fragment
     android:id="@+id/define_event"
     android:label="@string/event_photos"
-    android:name="edu.cnm.deepdive.picmegallery.controller.DefineEventFragment"
+    android:name="edu.cnm.deepdive.picmegallery.ui.defineEvent.DefineEventFragment"
     tools:layout="@layout/fragment_define_event"/>
 
 </navigation>
